separate components by domain
