+++ src/components/Experience/Markdown/renderers/AnchorBlock.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
 
-import { Color } from "../../../constants/Color";
+import { Color } from "../../../../constants/Color";
 
 export function AnchorBlock(props: any) {
   return props.href.startsWith("#") || props.href.startsWith("http") ? (
