+++ src/components/Project/ProjectSection.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import { TransitionContainer } from "../Layout/TransitionContainer";
+import { TransitionContainer } from "../common/TransitionContainer";
 import { ProjectCardList } from "./ProjectCardList";
 
 import { IProject } from "../../interfaces/section/Project";
