+++ src/components/Skill/SkillSection.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import { TransitionContainer } from "../Layout/TransitionContainer";
+import { TransitionContainer } from "../common/TransitionContainer";
 import { SkillCardList } from "./SkillCardList";
 
 import { ISkill } from "../../interfaces/section/Skill";
