+++ src/components/Values/ValuesSection.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import { TransitionContainer } from "../Layout/TransitionContainer";
+import { TransitionContainer } from "../common/TransitionContainer";
 import { ValuesCard } from "./ValuesCard";
 
 import { IValues } from "../../interfaces/section/Values";
